requirements.txt
astroid==2.2.5
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1
